+++ libs/render_delegate/render_pass.cpp
@@ -949,16 +949,21 @@ void HdArnoldRenderPass::_Execute(const HdRenderPassStateSharedPtr& renderPassSt
                                     .c_str()};
                         } else {
                             // Querying the data format from USD, with a default value of color3f.
-                            const auto format = _GetOptionalSetting<TfToken>(
-                                renderVar.settings, _tokens->dataType, _GetTokenFromHdFormat(renderVar.format));
+                            // If we have arnold:format defined, we use its value for the format
+                            const TfToken hydraFormat = _GetOptionalSetting<TfToken>(renderVar.settings, _tokens->dataType, _GetTokenFromHdFormat(renderVar.format));
+                            const TfToken arnoldFormat = _GetOptionalSetting<TfToken>(renderVar.settings, TfToken("arnold:format"), TfToken(""));
+                            const TfToken format = arnoldFormat != TfToken("") ? arnoldFormat : hydraFormat;
                             const auto arnoldTypes = _GetArnoldAOVTypeFromTokenType(format);
                             const auto aovName = _CreateAOV(
                                 _renderDelegate, arnoldTypes, renderVar.name, renderVar.sourceType,
                                 renderVar.sourceName, customRenderVar.writer, customRenderVar.reader, lightPathExpressions,
                                 aovShaders);
+                            // Check if the AOV has a specific filter
+                            const auto arnoldAovFilterName = _GetOptionalSetting<std::string>(renderVar.settings, TfToken("arnold:filter"), "");
+                            AtNode *aovFilterNode = arnoldAovFilterName.empty() ? nullptr : _CreateFilter(_renderDelegate, renderVar.settings);
                             customRenderVar.output =
                                 AtString{TfStringPrintf(
-                                             "%s %s %s %s", aovName.c_str(), arnoldTypes.outputString, filterName,
+                                             "%s %s %s %s", aovName.c_str(), arnoldTypes.outputString, aovFilterNode ? AiNodeGetName(aovFilterNode) : filterName,
                                              customDriverName.c_str())
                                              .c_str()};
                         }