+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug fixes
 - [usd#1538](https://github.com/Autodesk/arnold-usd/issues/1538) - Fix triplanar in USD Materialx
 - [usd#1588](https://github.com/Autodesk/arnold-usd/issues/1588) - Arnold schemas under a point instancer should be hidden
+- [usd#1595](https://github.com/Autodesk/arnold-usd/issues/1595) - Support Arnold RenderVar filters in Hydra 
 
 ## [7.2.2.1] - 2023-06-21